leancrypto: Certificates are now validated by BC
The resolution of one issue is quite interesting: I had to eliminate the O,
the ST and C field in the X.509 certs to make them pass. Interestingly,
OpenSSL (apart from the crypto aspect) can nicely parse my X.509 cert with
all name components.

In addition, leancrypto did not encode the pubkey for the
Composite-ML-DSA certs as a BIT STRING. The latest bug fix to leancrypto
fixed it.

Signed-off-by: Stephan Mueller <smueller@chronox.de>